# backend/openapi.yaml
openapi: 3.0.3
info:
  title: CyberOpus API
  description: |
    CyberOpus Online Bookstore REST API.

    ## Authentication
    Most endpoints require a JWT Bearer token obtained from `/api/auth/login` or `/api/auth/register`.
    Include it as: `Authorization: Bearer <token>`

    ## Test Payment
    Use `cardHolderName: "FAIL_TEST"` to simulate a payment failure.
    Normal names simulate success.
  version: "1.0"
  contact:
    name: CyberOpus
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderItemResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        bookId: { type: integer, format: int64 }
        bookTitle: { type: string }
        bookAuthor: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        totalPrice: { type: number }
    PaymentResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string, enum: [PENDING, SUCCESS, FAILED, REFUNDED] }
        cardHolderName: { type: string }
        maskedCardNumber: { type: string, example: "****4242" }
        amount: { type: number }
        processedAt: { type: string, format: date-time }
    OrderResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        status: { type: string, enum: [PENDING, CONFIRMED, CANCELLED] }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItemResponse' }
        subtotal: { type: number }
        deliveryCharge: { type: number }
        discount: { type: number }
        total: { type: number }
        rewardPointsEarned: { type: integer }
        rewardPointsRedeemed: { type: integer }
        snapshotFullName: { type: string }
        snapshotLine1: { type: string }
        snapshotCity: { type: string }
        snapshotState: { type: string }
        snapshotPostalCode: { type: string }
        snapshotCountry: { type: string }
        snapshotPhone: { type: string }
        placedAt: { type: string, format: date-time }
        canCancel: { type: boolean }
        payment: { $ref: '#/components/schemas/PaymentResponse' }
...
paths:
...
  /api/categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Category
    put:
      tags: [Categories, Admin]
      summary: Update category (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Updated category
    delete:
      tags: [Categories, Admin]
      summary: Delete category (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204':
          description: Deleted
  /api/categories/{id}/books:
    get:
      tags: [Categories]
      summary: Get books in category
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Books in category
...
  /api/orders:
    get:
      tags: [Orders]
      summary: List user's orders
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 10 }
      responses:
        '200':
          description: Paginated orders
  /api/orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderResponse' }
        '404':
          description: Order not found
        '403':
          description: Not your order
  /api/orders/{id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an order
      description: |
        Cancels a PENDING or CONFIRMED order within 48 hours of placement.
        Restores stock, reverses reward points, refunds payment.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Order cancelled
        '400':
          description: Cannot cancel (wrong status or past 48h window)
  /api/orders/{orderId}/buy-again:
    post:
      tags: [Orders]
      summary: Add order items back to cart
      description: Adds available books from a previous order to the cart at current prices.
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Items added to cart
  /api/payments/order/{orderId}:
    get:
      tags: [Payments]
      summary: Get payment details for an order
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentResponse' }
        '403':
          description: Not your payment
...